# contracts/vmm/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aseman VMM API",
    "version": "1.0.0",
    "description": "A501: the only interface between an Aseman node and a VMM (plan 04). States and generation rules are in `states.json` (A502, A503). Every request is authenticated by mutual TLS: the client certificate names the calling node; the VMM serves only that node's workloads."
  },
  "servers": [
    {
      "url": "https://vmm.internal"
    }
  ],
  "security": [
    {
      "mtls": []
    }
  ],
  "paths": {
    "/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Runtimes and operations this VMM supports",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/workloads": {
      "post": {
        "operationId": "createWorkload",
        "summary": "Create a workload with its desired state",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "description": "Refused with `unsupported_operation` when the runtime lacks what the spec needs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/Accepted"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "get": {
        "operationId": "listWorkloads",
        "summary": "List workloads",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "creature_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "observed_state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ObservedState"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadPage"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/workloads/{workload_id}": {
      "get": {
        "operationId": "getWorkload",
        "summary": "Read a workload's desired and observed state",
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkloadId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workload"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "The quoted resource version."
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "delete": {
        "operationId": "deleteWorkload",
        "summary": "Delete the workload (desired `deleted`)",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/WorkloadId"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "name": "generation",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Generation"
            }
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "responses": {
          "202": {
            "$ref": "#/components/responses/Accepted"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/workloads/{workload_id}/spec": {
      "put": {
        "operationId": "updateWorkloadSpec",
        "summary": "Replace the spec under a new generation",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/WorkloadId"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpec"
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/Accepted"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/workloads/{workload_id}/invocations": {
      "post": {
        "operationId": "invokeWorkload",
        "summary": "Deliver one event to the workload",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/WorkloadId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invocation"
              }
            }
          }
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/Accepted"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/workloads/{workload_id}/http": {
      "post": {
        "operationId": "forwardHttp",
        "summary": "Forward one HTTP request to the workload",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/WorkloadId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "description": "Synchronous within the deadline. The idempotency key deduplicates retried deliveries.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/operations/{operation_id}": {
      "get": {
        "operationId": "getOperation",
        "summary": "Read an operation",
        "parameters": [
          {
            "$ref": "#/components/parameters/OperationId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/Deadline"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mtls": {
        "type": "mutualTLS"
      }
    },
    "schemas": {
...
      "ObservedState": {
        "type": "string",
        "enum": [
          "unknown",
          "pending",
          "running",
          "paused",
          "stopped",
          "failed",
          "lost"
        ]
      },
...
      "Generation": {
        "type": "integer",
        "format": "int64",
        "minimum": 1,
        "description": "The node's desired generation (A503)."
      },
      "ResourceVersion": {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{1,64}$",
        "description": "Opaque; changes on every change to the record, desired or observed. Sent as a quoted ETag."
      },
...
      "WorkloadLabels": {
        "type": "object",
        "properties": {
          "creature_id": {
            "type": "string",
            "format": "uuid"
          },
          "program_id": {
            "type": "string",
            "format": "uuid"
          },
          "entity_id": {
            "type": "string",
            "maxLength": 128
          },
          "legacy_machine_id": {
            "type": "string",
            "maxLength": 256
          },
          "legacy_vm_id": {
            "type": "string",
            "maxLength": 256
          }
        },
        "additionalProperties": false,
        "required": [
          "creature_id",
          "entity_id",
          "program_id"
        ],
        "description": "Ownership labels set by the node. The VMM never derives authority from them."
      },
...
      "CreateWorkload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "labels": {
            "$ref": "#/components/schemas/WorkloadLabels"
          },
          "spec": {
            "$ref": "#/components/schemas/WorkloadSpec"
          },
          "desired": {
            "$ref": "#/components/schemas/DesiredStatus"
          }
        },
        "additionalProperties": false,
        "required": [
          "desired",
          "id",
          "labels",
          "spec"
        ],
        "description": "The node chooses the ID, so a retried create names the same workload."
      },
      "Workload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "labels": {
            "$ref": "#/components/schemas/WorkloadLabels"
          },
          "spec": {
            "$ref": "#/components/schemas/WorkloadSpec"
          },
          "desired": {
            "$ref": "#/components/schemas/DesiredStatus"
          },
          "applied_generation": {
            "$ref": "#/components/schemas/Generation"
          },
          "observed": {
            "$ref": "#/components/schemas/Observation"
          },
          "resource_version": {
            "$ref": "#/components/schemas/ResourceVersion"
          },
          "created_at_millis": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at_millis": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "required": [
          "created_at_millis",
          "desired",
          "id",
          "labels",
          "resource_version",
          "spec",
          "updated_at_millis"
        ]
      },
      "WorkloadPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workload"
            }
          },
          "next_cursor": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "items"
        ]
      },
...
      "UpdateSpec": {
        "type": "object",
        "properties": {
          "generation": {
            "$ref": "#/components/schemas/Generation"
          },
          "spec": {
            "$ref": "#/components/schemas/WorkloadSpec"
          }
        },
        "additionalProperties": false,
        "required": [
          "generation",
          "spec"
        ]
      },
      "Invocation": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "signal",
              "chain_transactions",
              "chain_effects"
            ]
          },
          "key": {
            "type": "string",
            "maxLength": 256,
            "description": "The signal key the program handles."
          },
          "store_id": {
            "type": "string",
            "maxLength": 256
          },
          "payload": {
            "description": "Any JSON value, at most 4 MiB encoded."
          }
        },
        "additionalProperties": false,
        "required": [
          "key",
          "kind",
          "payload"
        ],
        "description": "One event for the workload: runs the program once (invocation runtimes) or is delivered to the running instance."
      },
...
      "HttpRequest": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "pattern": "^[A-Z]{3,10}$"
          },
          "path": {
            "type": "string",
            "maxLength": 8192
          },
          "query": {
            "type": "string",
            "maxLength": 8192
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "maxLength": 8192
            },
            "maxProperties": 128
          },
          "body": {
            "type": "string",
            "contentEncoding": "base64url",
            "description": "Unpadded base64url (RFC 4648 section 5)."
          },
          "port": {
            "type": "string",
            "description": "Ingress port name; the first `http` port when absent."
          }
        },
        "additionalProperties": false,
        "required": [
          "headers",
          "method",
          "path"
        ]
      },
      "HttpResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "body": {
            "type": "string",
            "contentEncoding": "base64url",
            "description": "Unpadded base64url (RFC 4648 section 5)."
          }
        },
        "additionalProperties": false,
        "required": [
          "headers",
          "status"
        ]
      },
...
      "Capabilities": {
        "type": "object",
        "properties": {
          "api_version": {
            "type": "string"
          },
          "backend": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "version": {
                "type": "string"
              }
            },
            "additionalProperties": false,
            "required": [
              "name",
              "version"
            ]
          },
          "runtimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuntimeCapabilities"
            }
          },
          "max_request_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false,
        "required": [
          "api_version",
          "backend",
          "max_request_bytes",
          "runtimes"
        ]
      },
...
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Required on every mutation. A replay with the same key and body returns the first response; the same key with another body is `idempotency_key_reused`. Keys are kept for 24 hours.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{16,128}$"
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": false,
        "description": "Optional quoted resource version; a mismatch is `resource_version_mismatch` (412).",
        "schema": {
          "type": "string"
        }
      },
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Echoed in responses, problems, operations, and logs; generated when absent.",
        "schema": {
          "type": "string",
          "maxLength": 128
        }
      },
      "Traceparent": {
        "name": "traceparent",
        "in": "header",
        "required": false,
        "description": "W3C Trace Context, propagated to the backend.",
        "schema": {
          "type": "string"
        }
      },
      "Deadline": {
        "name": "Aseman-Deadline",
        "in": "header",
        "required": false,
        "description": "Unix milliseconds after which the caller no longer wants the result; the VMM gives up with `deadline_exceeded` and cancels the operation.",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
      "WorkloadId": {
        "name": "workload_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "OperationId": {
        "name": "operation_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 512
        },
        "description": "Opaque `next_cursor` from the previous page."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "A problem (RFC 9457).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Accepted": {
        "description": "The operation was recorded; poll it or follow the workload's events.",
        "headers": {
          "Location": {
            "schema": {
              "type": "string"
            },
            "description": "`/v1/operations/{operation_id}`."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Operation"
            }
          }
        }
      }
    }
  }
}